:fire: remove utils.py
